pyproject.toml
[tool.poetry]
name = "bcraft-recipes"
version = "0.1.0"
description = "Test task from Bcraft company"
authors = ["Gleb <maxwellgleb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">= 3.10, < 3.12"
SQLAlchemy = {extras = ["asyncio"], version = "~2.0.10"}
pydantic = {extras = ["dotenv"], version = "~1.10.7"}
alembic = "~1.10.4"
fastapi = {extras = ["all"], version = "~0.95.1"}
fastapi-pagination = "~0.12.2"
asyncpg = "~0.27.0"
fastapi-users = {extras = ["sqlalchemy"], version = "~10.4.2"}
Pillow = "~9.5.0"
aiofiles = "~23.1.0"
Faker = "~18.6.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"
types-aiofiles = "^23.1.0.1"
types-Pillow = "^9.5.0.2"
black = "^23.3.0"
SQLAlchemy-Utils = "^0.41.1"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.0.0"
pytest-check = "^2.1.4"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
